+++ tests/Aer.QdrantClient.Tests/TestClasses/HttpClientTests/Cluster/ClusterCompoundOperationsTests.cs
